chore(release): version packages (#195)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @folks-finance/xchain-sdk@0.0.42

### Patch Changes

-
[#194](https://github.com/Folks-Finance/folks-finance-xchain-js-sdk/pull/194)
[`decfce7`](https://github.com/Folks-Finance/folks-finance-xchain-js-sdk/commit/decfce70c82ad01703799f6af21c1687b61bb70f)
Thanks [@palace22](https://github.com/palace22)! - Fixed check on cross
chain token adapter when the chain is the hub

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>